+++ src/lib/Scenes/Home/__stories__/Home.story.tsx
@@ -1,16 +1,11 @@
 import { storiesOf } from "@storybook/react-native"
-import React from "react"
-
-import Home from "../"
-
 import WorksForYou from "lib/Containers/WorksForYou"
-
-import ForYou from "../Components/ForYou"
-import Sales from "../Components/Sales"
-
 import { ForYouRenderer, WorksForYouRenderer } from "lib/relay/QueryRenderers"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
-import SalesRenderer from "../Components/Sales/Relay/SalesRenderer"
+import React from "react"
+import Home from "../"
+import ForYou from "../Components/ForYou"
+import Sales, { SalesRenderer } from "../Components/Sales"
 
 storiesOf("Home/Relay")
   .add("Root", () => <Home />)