+++ CHANGELOG.md
@@ -8,12 +8,14 @@
 -   Don’t assume a photo was selected in the consignments photo picker - alloy
 -   QA for composer/thread view: ipad and fixed iphone size attachments - maxim
 -   QA for message attachments - maxim
+-   QA for saves & follows: enable pagination in Works tab in Saves & Follows - luc
 -   QA for home header font sizes and margins - maxim
 -   Minor: change text colour for provenance placeholder in consignments - maxim
 -   Don’t rely on user of the Gene VC to specify filter defaults, which could lead to a crash - alloy
 -   QA for inbox zero on ipad - maxim
 -   Add analytics to consignments - orta
 -   Fixes for metaphysics changes to consignment submission - orta
+-   Implement selected state for images in consignments photo set - maxim
 
 ### 1.4.0-beta.8
 
@@ -82,6 +84,7 @@
 -   Adds Live and timed auctions in home view - luc
 -   Tab bar text now aligned and centered correctly - luc
 -   Adds iPad support for Auctions view layout - luc
+-   Adds <LotsByArtistsYouFollow /> component to Sales - chris
 
 ###### Consignments